+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SaintsGraph.Editor")]
+[assembly: InternalsVisibleTo("SaintsGraph.Tests.Editor")]